refactored one dirname
